requirements.txt
Flask==3.0.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
Werkzeug==3.0.1